+++ app/app/models/category.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, Float, Integer, String
+from sqlalchemy import Boolean, Column, Numeric, Integer, String
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql.expression import false
 from app.db.base_class import Base
@@ -11,7 +11,7 @@ class Category(Base):
     id = Column(Integer, primary_key=True, index=True)
     title = Column(String, index=True)
     borrow_limit = Column(Integer, nullable=False)
-    borrow_price_per_day = Column(Float, nullable=False)
+    borrow_price_per_day = Column(Numeric(10, 2), nullable=False)
     is_deleted = Column(Boolean(), nullable=False, default=False)
 
     books = relationship('Book', back_populates="category")
